Part that was already done
The lay out of the website was done. 

In the 'relaties' page there is a table showing the contents of the database. The modal displaying the information of these database entries is only working if redirected to via a normal link. If the whole line of the table is clickable, it doesn't display the full information.

There also is an add button that redirects to a form (not working yet). In the comments of the 404 page there is a form that has direct confirmation of the boxes in the form, but somehow that didn't work properly in the add page (might have something to do with the modal it was earlier placed in.